+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableCharArray.kt
@@ -8,6 +8,7 @@ import kotlin.Array
 import kotlin.Boolean
 import kotlin.Char
 import kotlin.CharArray
+import kotlin.Comparable
 import kotlin.Int
 import kotlin.PublishedApi
 import kotlin.String
@@ -215,7 +216,17 @@ public value class ImmutableCharArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
-     * Leaves [this] immutable array as is and returns an [ImmutableCharArray] with all elements
+     * Leaves this immutable array as is and returns an ImmutableCharArray with all elements sorted
+     * according to the natural sort order of the value returned by specified [selector].
+     *
+     * The sort is _stable_ so equal elements preserve their order relative to each other after
+     * sorting.
+     */
+    public inline fun <R : Comparable<R>> sortedBy(crossinline selector: (element: Char) -> R?):
+            ImmutableCharArray = sortedWith(compareBy(selector))
+
+    /**
+     * Leaves this immutable array as is and returns an [ImmutableCharArray] with all elements
      * sorted according to the specified [comparator].
      */
     public fun sortedWith(comparator: Comparator<in Char>): ImmutableCharArray {