+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableDoubleArray.kt
@@ -6,6 +6,7 @@ import java.util.Arrays
 import java.util.Comparator
 import kotlin.Array
 import kotlin.Boolean
+import kotlin.Comparable
 import kotlin.Double
 import kotlin.DoubleArray
 import kotlin.Int
@@ -217,7 +218,17 @@ public value class ImmutableDoubleArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
-     * Leaves [this] immutable array as is and returns an [ImmutableDoubleArray] with all elements
+     * Leaves this immutable array as is and returns an ImmutableDoubleArray with all elements
+     * sorted according to the natural sort order of the value returned by specified [selector].
+     *
+     * The sort is _stable_ so equal elements preserve their order relative to each other after
+     * sorting.
+     */
+    public inline fun <R : Comparable<R>> sortedBy(crossinline selector: (element: Double) -> R?):
+            ImmutableDoubleArray = sortedWith(compareBy(selector))
+
+    /**
+     * Leaves this immutable array as is and returns an [ImmutableDoubleArray] with all elements
      * sorted according to the specified [comparator].
      */
     public fun sortedWith(comparator: Comparator<in Double>): ImmutableDoubleArray {