+++ immutable-arrays/core/src/test/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArrayTest.kt
@@ -460,6 +460,18 @@ class ImmutableArrayTest {
     }
 
     @Test
+    fun `sortedBy validation`() {
+        with(immutableArrayOf("apples", "pineapples", "dogs")) {
+            expectThat(sortedBy { it.length })
+                .isEqualTo(immutableArrayOf("dogs", "apples", "pineapples"))
+
+            // original is left unchanged
+            expectThat(this)
+                .isEqualTo(immutableArrayOf("apples", "pineapples", "dogs"))
+        }
+    }
+
+    @Test
     fun `sortedWith validation`() {
         with(immutableArrayOf("dogs", "apples", "pineapples")) {
             val comparator = Comparator<String> { o1, o2 -> o1.compareTo(o2) }